+++ backend/src/models/sales_quotation.rs
@@ -10,7 +10,7 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i64,
     pub quotation_no: String,
-    pub customer_id: i32,
+    pub customer_id: i64,
     pub sales_user_id: i64,
     pub quotation_date: NaiveDate,
     pub valid_until: NaiveDate,
@@ -50,7 +50,7 @@ pub struct Model {
 
     /// BPM 审批
     pub approval_instance_id: Option<i64>,
-    pub approved_by: Option<i32>,
+    pub approved_by: Option<i64>,
     pub approved_at: Option<DateTime<Utc>>,
     pub rejection_reason: Option<String>,
